requirements.txt
# Polymarket Market Making Bot – Python dependencies
# Install with:  pip install -r requirements.txt

# Official Polymarket Python SDK (CLOB REST + order signing)
py-clob-client>=0.19.0

# Async WebSocket client (order book feed)
websockets>=12.0

# Synchronous HTTP (Gamma API + fallback REST calls)
requests>=2.31.0

# Async HTTP (optional: swap out requests in future)
aiohttp>=3.9.0

# Blockchain interaction (position merging on Polygon)
web3>=6.15.0

# Environment variable management
python-dotenv>=1.0.0

# Structured logging (optional: drop-in for standard logging)
loguru>=0.7.0

# Data handling (optional: for analysis / backtesting scripts)
pandas>=2.0.0
